Convert jest to vitest
