+++ app/client/.eslintignore
@@ -1,6 +1,8 @@
 tailwind.config.js
-craco.config.js
 node_modules
 cypress
 cypress.config.js
 build
+env.d.ts
+dist/
+.eslintrc.js
