+++ app/client/package.json
@@ -25,20 +25,25 @@
     "start": "vite",
     "build": "tsc && vite build",
     "serve": "vite preview",
-    "test": "vite test",
-    "prettier": "prettier --write ."
+    "test": "vitest watch",
+    "test:no-watch": "vitest run",
+    "test:coverage": "vitest run --coverage",
+    "lint": "eslint . --ext .ts,.tsx"
   },
   "license": "UNLICENSED",
   "devDependencies": {
     "@babel/core": "^7.13.0",
     "@tanstack/react-query-devtools": "^4.23.0",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@types/jest": "^29.2.6",
     "@typescript-eslint/eslint-plugin": "^5.41.0",
     "@typescript-eslint/parser": "^5.41.0",
     "@vitejs/plugin-react": "^3.0.1",
+    "@vitest/coverage-c8": "^0.28.1",
     "autoprefixer": "^10.4.13",
     "c8": "^7.12.0",
+    "cross-fetch": "^3.1.5",
     "cypress": "^10.5.0",
     "eslint": "^8.26.0",
     "eslint-config-airbnb": "^19.0.4",
@@ -51,6 +56,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "jest": "^29.3.1",
     "jsdom": "^21.1.0",
+    "msw": "^0.49.3",
     "nock": "^13.3.0",
     "postcss": "^8.4.21",
     "prettier": "^2.2.1",
