+++ app/client/src/mocks/server.ts
@@ -1,0 +1,4 @@
+import { setupServer } from 'msw/node'
+import { handlers } from './handler'
+// This configures a Service Worker with the given request handlers.
+export const server = setupServer(...handlers)
