+++ app/client/tsconfig.json
@@ -3,7 +3,7 @@
         "module": "commonjs",
         "target": "ESNext",
         "lib": ["dom", "dom.iterable", "esnext"],
-        "types": ["vite/client", "vite-plugin-svgr/client"],
+        "types": ["vite/client", "vite-plugin-svgr/client","vitest/globals"],
         "allowJs": true,
         "checkJs": false,
         "jsx": "react",
